[Refactoring] CmdBuffer Class화, Parameterized UnitTest Case 환경 Setting
